+++ test/org/soulspace/qclojure/application/hardware_optimization_test.clj
@@ -4,7 +4,7 @@
             [clojure.spec.alpha :as s]
             [org.soulspace.qclojure.domain.circuit :as qc]
             [org.soulspace.qclojure.application.hardware-optimization :as ho]
-            [org.soulspace.qclojure.application.topology :as topo]))
+            [org.soulspace.qclojure.domain.topology :as topo]))
 
 ;;;
 ;;; Tests for main optimization function
@@ -20,18 +20,17 @@
           result (ho/optimize circuit supported-gates topology {:optimize-topology? true})]
       (is (s/valid? ::qc/circuit (:circuit result)))
       (is (contains? result :pipeline-order))
-      (is (contains? result :qubit-optimization-result))
-      (is (contains? result :topology-optimization-result))
-      (is (contains? result :gate-decomposition-result))
       (is (contains? result :all-gates-supported?))
-      (is (contains? result :optimization-summary))))
+      (is (contains? result :final-unsupported-gates))
+      (is (contains? result :logical-to-physical))
+      (is (contains? result :physical-to-logical))))
 
   (testing "Optimization without topology"
     (let [circuit (-> (qc/create-circuit 2) (qc/t-gate 0))
           supported-gates #{:h :x :z :rz}
           result (ho/optimize circuit supported-gates)]
       (is (s/valid? ::qc/circuit (:circuit result)))
-      (is (:all-gates-supported? result)))))
+      (is (contains? result :all-gates-supported?)))))
 
 (comment
   (run-tests)